WIP versioning support
Clients (api-server & seasketch) should respect versioning of
reports following a semver-based scheme. For example, if SeaSketch
is on 2.0.1 of the report environment but the report was deployed
with 1.4.2, the report won't display. 2.1.0 vs 2.0.5-report would
work fine though.

Right now the system is in an unbuildable state due to more module
resolution issues... sigh. I'll need to break out the legacy-shim
app into it's own package to resolve these issues.

References #49. Closes #55.